refactor(database): emit UserTOTP SETs without coalesce
Drop the patch/coalesce helpers so each typed update writes one SET.
Use uint8 for increment Delta so negatives are a compile error, and
let duplicate column assignments fail at the database.

+++ internal/domain/user_totp_update.go
@@ -27,7 +27,7 @@ type UserTOTPLastSuccessfulCheckUpdate struct {
 func (*UserTOTPLastSuccessfulCheckUpdate) userTOTPUpdate() {}
 
 type UserTOTPIncrementFailedAttemptsUpdate struct {
-	Delta int16
+	Delta uint8
 }
 
 func (*UserTOTPIncrementFailedAttemptsUpdate) userTOTPUpdate() {}
